requirements.txt
streamlit
lark-parser
graphviz
pydot
